# docs/openapi-semana9.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MiniMarket Plus API",
    "description": "API REST de MiniMarket Plus para autenticación, catálogo, carritos, inventario, ventas, sucursales, stock por sucursal, proveedores, órdenes de compra, pedidos, promociones y reportes. Protegida mediante JWT y enriquecida con enlaces HATEOAS.",
    "contact": {
      "name": "Equipo MiniMarket Plus"
    },
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Servidor local"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/sucursales/{id}": {
      "get": {
        "tags": [
          "Sucursales"
        ],
        "operationId": "obtener",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la sucursal",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Sucursal encontrada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelSucursal"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          },
          "404": {
            "description": "Sucursal no encontrada"
          }
        },
        "summary": "Obtener sucursal",
        "description": "Obtiene una sucursal por su identificador."
      },
      "put": {
        "tags": [
          "Sucursales"
        ],
        "operationId": "actualizar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la sucursal",
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sucursal"
              }
            }
          },
          "required": true,
          "description": "Datos actualizados de la sucursal"
        },
        "responses": {
          "200": {
            "description": "Sucursal actualizada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelSucursal"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Sucursal no encontrada"
          },
          "409": {
            "description": "Nombre de sucursal duplicado"
          }
        },
        "summary": "Actualizar sucursal",
        "description": "Actualiza los datos de una sucursal existente. Requiere ROLE_ADMIN."
      },
      "delete": {
        "tags": [
          "Sucursales"
        ],
        "operationId": "eliminar",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la sucursal",
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Sucursal eliminada"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Sucursal no encontrada"
          },
          "409": {
            "description": "La sucursal posee relaciones activas"
          }
        },
        "summary": "Eliminar sucursal",
        "description": "Elimina una sucursal existente. Requiere ROLE_ADMIN."
      }
    },
    "/api/proveedores/{id}": {
      "get": {
        "tags": [
          "Proveedores"
        ],
        "operationId": "obtener_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del proveedor",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Proveedor encontrado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProveedor"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Proveedor no encontrado"
          }
        },
        "summary": "Obtener proveedor",
        "description": "Obtiene un proveedor por su identificador. Requiere ROLE_ADMIN."
      },
      "put": {
        "tags": [
          "Proveedores"
        ],
        "operationId": "actualizar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del proveedor",
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Proveedor"
              }
            }
          },
          "required": true,
          "description": "Datos actualizados del proveedor"
        },
        "responses": {
          "200": {
            "description": "Proveedor actualizado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelProveedor"
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Proveedor no encontrado"
          },
          "409": {
            "description": "Correo de proveedor duplicado"
          }
        },
        "summary": "Actualizar proveedor",
        "description": "Actualiza un proveedor existente. Requiere ROLE_ADMIN."
      },
      "delete": {
        "tags": [
          "Proveedores"
        ],
        "operationId": "eliminar_1",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del proveedor",
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Proveedor eliminado"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Proveedor no encontrado"
          },
          "409": {
            "description": "El proveedor posee órdenes relacionadas"
          }
        },
        "summary": "Eliminar proveedor",
        "description": "Elimina un proveedor existente. Requiere ROLE_ADMIN."
      }
    },
    "/api/promociones/{id}": {
      "get": {
        "tags": [
          "Promociones"
        ],
        "operationId": "obtener_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la promoción",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Promoción encontrada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPromocion"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Promoción no encontrada"
          }
        },
        "summary": "Obtener promoción",
        "description": "Obtiene una promoción por su identificador. Requiere ROLE_ADMIN."
      },
      "put": {
        "tags": [
          "Promociones"
        ],
        "operationId": "actualizar_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la promoción",
            "example": 1
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromocionRequest"
              }
            }
          },
          "required": true,
          "description": "Datos actualizados de la promoción"
        },
        "responses": {
          "200": {
            "description": "Promoción actualizada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPromocion"
                }
              }
            }
          },
          "400": {
            "description": "Datos o fechas inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Promoción o producto no encontrado"
          }
        },
        "summary": "Actualizar promoción",
        "description": "Actualiza una promoción existente. Requiere ROLE_ADMIN."
      },
      "delete": {
        "tags": [
          "Promociones"
        ],
        "operationId": "eliminar_2",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la promoción",
            "example": 1
          }
        ],
        "responses": {
          "204": {
            "description": "Promoción eliminada"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Promoción no encontrada"
          }
        },
        "summary": "Eliminar promoción",
        "description": "Elimina una promoción existente. Requiere ROLE_ADMIN."
      }
    },
...
    "/api/sucursales": {
      "get": {
        "tags": [
          "Sucursales"
        ],
        "summary": "Listar sucursales",
        "operationId": "listar",
        "responses": {
          "200": {
            "description": "Sucursales obtenidas correctamente",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelSucursal"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          }
        },
        "description": "Obtiene las sucursales registradas y enlaces HATEOAS hacia su stock."
      },
      "post": {
        "tags": [
          "Sucursales"
        ],
        "operationId": "crear",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Sucursal"
              }
            }
          },
          "required": true,
          "description": "Datos de la sucursal"
        },
        "responses": {
          "201": {
            "description": "Sucursal creada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelSucursal"
                }
              }
            }
          },
          "400": {
            "description": "Nombre o dirección inválidos"
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "409": {
            "description": "Nombre de sucursal duplicado"
          }
        },
        "summary": "Crear sucursal",
        "description": "Crea una sucursal. Requiere ROLE_ADMIN."
      }
    },
...
    "/api/stock-sucursales/sucursal/{sucursalId}": {
      "get": {
        "tags": [
          "Stock por sucursal"
        ],
        "operationId": "listarPorSucursal",
        "parameters": [
          {
            "name": "sucursalId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la sucursal",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Stock de la sucursal obtenido correctamente",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelStockSucursal"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          },
          "404": {
            "description": "Sucursal no encontrada"
          }
        },
        "summary": "Listar stock de una sucursal",
        "description": "Obtiene las existencias de todos los productos en una sucursal."
      }
    },
...
    "/api/promociones/activas": {
      "get": {
        "tags": [
          "Promociones"
        ],
        "operationId": "activas",
        "responses": {
          "200": {
            "description": "Promociones activas obtenidas correctamente",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/CollectionModelEntityModelPromocion"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          }
        },
        "summary": "Listar promociones activas",
        "description": "Obtiene las promociones vigentes para los usuarios autenticados."
      }
    },
    "/api/pedidos/{id}": {
      "get": {
        "tags": [
          "Pedidos en línea"
        ],
        "operationId": "obtener_3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador del pedido",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Pedido encontrado",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelPedido"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Rol no autorizado"
          },
          "404": {
            "description": "Pedido no encontrado"
          }
        },
        "summary": "Obtener pedido",
        "description": "ROLE_CLIENTE solo puede consultar pedidos propios; ROLE_ADMIN puede consultar cualquiera."
      }
    },
...
    "/api/ordenes-compra/{id}": {
      "get": {
        "tags": [
          "Órdenes de compra"
        ],
        "operationId": "obtener_4",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "description": "Identificador de la orden",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Orden encontrada",
            "content": {
              "application/hal+json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityModelOrdenCompra"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente o inválido"
          },
          "403": {
            "description": "Se requiere ROLE_ADMIN"
          },
          "404": {
            "description": "Orden no encontrada"
          }
        },
        "summary": "Obtener orden de compra",
        "description": "Obtiene una orden por su identificador. Requiere ROLE_ADMIN."
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Sucursal": {
        "type": "object",
        "description": "Sucursal de MiniMarket Plus.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "nombre": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "activa": {
            "type": "boolean"
          }
        }
      },
      "EntityModelSucursal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "nombre": {
            "type": "string"
          },
          "direccion": {
            "type": "string"
          },
          "activa": {
            "type": "boolean"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "Proveedor": {
        "type": "object",
        "description": "Proveedor utilizado para las órdenes de compra.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "nombre": {
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          }
        }
      },
      "EntityModelProveedor": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "nombre": {
            "type": "string"
          },
          "correo": {
            "type": "string"
          },
          "telefono": {
            "type": "string"
          },
          "activo": {
            "type": "boolean"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "PromocionRequest": {
        "type": "object",
        "description": "Datos de una promoción centralizada.",
        "properties": {
          "nombre": {
            "type": "string"
          },
          "productoId": {
            "type": "integer",
            "format": "int64"
          },
          "porcentajeDescuento": {
            "type": "number",
            "format": "double"
          },
          "fechaInicio": {
            "type": "string",
            "format": "date"
          },
          "fechaFin": {
            "type": "string",
            "format": "date"
          },
          "activa": {
            "type": "boolean"
          }
        }
      },
...
      "EntityModelPromocion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "nombre": {
            "type": "string"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          },
          "porcentajeDescuento": {
            "type": "number",
            "format": "double"
          },
          "fechaInicio": {
            "type": "string",
            "format": "date"
          },
          "fechaFin": {
            "type": "string",
            "format": "date"
          },
          "activa": {
            "type": "boolean"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "Producto": {
        "type": "object",
        "description": "Producto disponible en MiniMarket Plus.",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Identificador del producto",
            "example": 1,
            "readOnly": true
          },
          "nombre": {
            "type": "string",
            "description": "Nombre comercial",
            "example": "Arroz 1 kg"
          },
          "precio": {
            "type": "number",
            "format": "double",
            "description": "Precio unitario",
            "example": 1890.0
          },
          "stock": {
            "type": "integer",
            "format": "int32",
            "description": "Stock disponible",
            "example": 40
          },
          "categoria": {
            "$ref": "#/components/schemas/Categoria",
            "description": "Categoría asociada al producto"
          }
        }
      },
...
      "EntityModelPedido": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "usuario": {
            "$ref": "#/components/schemas/Usuario"
          },
          "sucursal": {
            "$ref": "#/components/schemas/Sucursal"
          },
          "fecha": {
            "type": "string",
            "format": "date-time"
          },
          "tipoEntrega": {
            "type": "string"
          },
          "direccionDespacho": {
            "type": "string"
          },
          "estado": {
            "type": "string"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "EntityModelStockSucursal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "sucursal": {
            "$ref": "#/components/schemas/Sucursal"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          },
          "stockActual": {
            "type": "integer",
            "format": "int32"
          },
          "stockMinimo": {
            "type": "integer",
            "format": "int32"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "EntityModelOrdenCompra": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "sucursal": {
            "$ref": "#/components/schemas/Sucursal"
          },
          "proveedor": {
            "$ref": "#/components/schemas/Proveedor"
          },
          "producto": {
            "$ref": "#/components/schemas/Producto"
          },
          "cantidad": {
            "type": "integer",
            "format": "int32"
          },
          "fechaCreacion": {
            "type": "string",
            "format": "date-time"
          },
          "estado": {
            "type": "string"
          },
          "automatica": {
            "type": "boolean"
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "CollectionModelEntityModelSucursal": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "sucursalList": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EntityModelSucursal"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
      "CollectionModelEntityModelStockSucursal": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "stockSucursalList": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EntityModelStockSucursal"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "CollectionModelEntityModelPromocion": {
        "type": "object",
        "properties": {
          "_embedded": {
            "type": "object",
            "properties": {
              "promocionList": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EntityModelPromocion"
                }
              }
            }
          },
          "_links": {
            "$ref": "#/components/schemas/Links"
          }
        }
      },
...
      "Link": {
        "type": "object",
        "properties": {
          "href": {
            "type": "string"
          },
          "hreflang": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "deprecation": {
            "type": "string"
          },
          "profile": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "templated": {
            "type": "boolean"
          }
        }
      },
      "Links": {
        "type": "object",
        "additionalProperties": {
          "$ref": "#/components/schemas/Link"
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "Ingrese únicamente el token generado por POST /api/auth/login.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}